pyproject.toml
[project]
name = "insight-capsule"
version = "0.2.0"
description = "A local-first, voice-driven thought capture tool with AI-powered synthesis"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Ryan Johnson", email = "ryan@ryanleej.com"}
]
keywords = ["ai", "voice", "transcription", "whisper", "ollama", "accessibility"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core AI/ML
    "openai>=1.80.0",
    "openai-whisper>=20231117",
    "numpy<2",  # Required for whisper compatibility

    # Audio Processing
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",

    # Text-to-Speech
    "pyttsx3>=2.90",

    # HTTP & Configuration
    "requests>=2.32.3",
    "python-dotenv>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

[project.scripts]
insight-capsule = "main:main"
insight-cli = "cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "agents", "pipeline", "config", "utils"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
# Exclude interactive manual tests from automated test runs
addopts = "-v --cov=core --cov=agents --cov=pipeline --ignore=tests/test_tts_minimal.py --ignore=tests/test_tts.py"
